clean up function signatures/removed functions remove unused functions convert warnings to errors resolve relevant warnings

+++ libnd4j/include/ops/declarable/generic/reduce/reduceMean.cpp
@@ -44,7 +44,7 @@ CUSTOM_OP_IMPL(reduce_mean, -1, 1, false, 0, 0) {
     keepDims = (bool)T_ARG(0);
 
   REQUIRE_TRUE(
-      dimensions.size() <= input->rankOf(), 0,
+      dimensions.size() <= static_cast<size_t>(input->rankOf()), 0,
       "REDUCE_MEAN OP: the number of dimensions to reduce along must be <= input array rank, but got %i instead",
       dimensions.size());
 
@@ -73,7 +73,7 @@ DECLARE_SHAPE_FN(reduce_mean) {
     keepDims = (bool)T_ARG(0);
 
   REQUIRE_TRUE(
-      dimensions.size() <= in[0], 0,
+      dimensions.size() <= static_cast<size_t>(in[0]), 0,
       "REDUCE_MEAN OP: the number of dimensions to reduce along must be <= input array rank, but got %i instead",
       dimensions.size());
 
@@ -112,7 +112,7 @@ CUSTOM_OP_IMPL(reduce_mean_bp, -2, 1, false, 0, 0) {
     keepDims = (bool)T_ARG(0);
 
   REQUIRE_TRUE(
-      dimensions.size() <= input->rankOf(), 0,
+      dimensions.size() <= static_cast<size_t>(input->rankOf()), 0,
       "REDUCE_MEAN_BP OP: the number of dimensions to reduce along must be <= input array rank, but got %i instead",
       dimensions.size());
 
@@ -162,7 +162,7 @@ DECLARE_SHAPE_FN(reduce_mean_bp) {
     helpers::adjustAxis(rank, axesVector, dimensions);
   }
   REQUIRE_TRUE(
-      dimensions.size() <= rank, 0,
+      dimensions.size() <= static_cast<size_t>(rank), 0,
       "REDUCE_MEAN_BP OP: the number of dimensions to reduce along must be <= input array rank, but got %i instead",
       dimensions.size());
